game_dir: glsl: fix gamma-correction in vertex lighting (#225)
Apply gamma-correction before summing lightstyles.
A better way would be to do things like gamma-correction and HDR decompression on the CPU and upload per-vertex light and deluxe values to GPU as half floats.
This would simplify the shader. And UnpackVector and UnpackNormal are not very precise.
Another possible improvement would be to not sum lightstyles in the vertex shaders but to pass all the values to the fragment shader and calculate each lightstyle separately.